refactoring #99
